refactor: consolidate CRM entity interfaces into CrmTypes.ts for better organization
