+++ frontend/src/community/crm/types/CrmCommonTypes.ts
@@ -1,159 +1,17 @@
 import { SortOrderTypes } from "~community/common/types/CommonTypes";
 
+import { CrmDealSortEnum, CrmPriorityEnum } from "../enums/common";
 import {
-  CrmDealSortEnum,
-  CrmDealStageColorsEnum,
-  CrmDealStageEnum,
-  CrmIndustryEnum,
-  CrmPriorityEnum
-} from "../enums/common";
-
-// Company
-
-export interface CrmCompanyEntity {
-  id?: number;
-  name?: string;
-  industry?: CrmIndustryEnum;
-  website?: string;
-  address?: string;
-  contactNumber?: string;
-  openTasksCount?: number;
-  overdue?: number;
-  openValue?: string;
-  accountValue?: string;
-  openDeals?: number;
-  closedDeals?: number;
-  contactIds?: number[];
-  dealIds?: number[];
-  taskIds?: number[];
-}
-
-export interface CrmCompanyMetrics {
-  id?: number;
-  openTasksCount?: number;
-  overdue?: number;
-  openValue?: string;
-  accountValue?: string;
-  openDeals?: number;
-  closedDeals?: number;
-}
-
-export interface CrmCompanyLookup {
-  id: number;
-  name: string;
-}
-
-// Contact
-
-export interface CrmContactEntity {
-  id?: number;
-  firstName?: string;
-  lastName?: string;
-  email?: string;
-  contactNumber?: string;
-  lastContactAt?: string;
-  lastModifiedDate?: string;
-  companyId?: number;
-  ownerId?: number;
-  totalRevenue?: string;
-  pipelineRevenue?: string;
-  activeDealsCount?: number;
-  openTasksCount?: number;
-  overdueTasksCount?: number;
-  closedDealValue?: string;
-  closedDealCount?: number;
-  dealIds?: number[];
-  taskIds?: number[];
-}
-
-export interface CrmContactMetrics {
-  id?: number;
-  totalRevenue?: string;
-  pipelineRevenue?: string;
-  activeDealsCount?: number;
-  openTasksCount?: number;
-  overdueTasksCount?: number;
-  closedDealValue?: string;
-  closedDealCount?: number;
-}
-
-export interface CrmContactLookupItem {
-  id: number;
-  firstName: string;
-  lastName?: string;
-  company?: CrmCompanyLookup;
-}
-
-export interface CrmOwnerEntity {
-  employeeId: number;
-  firstName: string;
-  lastName?: string;
-  email?: string;
-  authPic?: string;
-}
-
-// Deal
-
-export interface CrmDealEntity {
-  id?: number;
-  name?: string;
-  description?: string;
-  priority?: CrmPriorityEnum;
-  orderIndex?: string;
-  amount?: string;
-  closingAt?: string;
-  stageId?: number;
-  companyId?: number;
-  contactId?: number;
-  ownerId?: number;
-  openTasksCount?: number;
-  taskIds?: number[];
-}
-
-export interface CrmDealLookupItem {
-  id: number;
-  name: string;
-}
-
-export interface CrmDealStageEntity {
-  id?: number;
-  name?: string;
-  description?: string;
-  color?: CrmDealStageColorsEnum;
-  orderIndex?: number;
-  stageType?: CrmDealStageEnum;
-}
-
-export interface CrmBoardColumn {
-  dealIds: number[];
-  totalCount: number;
-  currentPage: number;
-  hasNextPage: boolean;
-}
-
-// Task — entity and its type
-
-export interface CrmTaskEntity {
-  id?: number;
-  name?: string;
-  priority?: CrmPriorityEnum;
-  isCompleted?: boolean;
-  dueAt?: string;
-  notes?: string;
-  typeId?: number;
-  ownerId?: number;
-  contactId?: number;
-  companyId?: number;
-  dealId?: number;
-}
-
-export interface CrmTaskType {
-  id: number;
-  name: string;
-  orderIndex: number;
-}
+  CrmCompanyEntity,
+  CrmContactEntity,
+  CrmDealEntity,
+  CrmOwnerEntity,
+  CrmTaskEntity,
+  CrmTaskTypeEntity
+} from "./CrmTypes";
 
 // Requests
+
 export interface CrmDealReorderWithinStageRequest {
   dealId: number;
   previousDealId: number | null;
@@ -173,15 +31,9 @@ export interface CrmDealStageReorderRequest {
 }
 
 // Responses
+
 export interface CrmCompanyListResponse {
   items: CrmCompanyEntity[];
-  currentPage: number;
-  totalItems: number;
-  totalPages: number;
-}
-
-export interface CrmCompanyLookupListResponse {
-  items: CrmCompanyLookup[];
   currentPage: number;
   totalItems: number;
   totalPages: number;
@@ -193,13 +45,6 @@ export interface CrmCompanyDomainSearchResponse {
 
 export interface CrmContactListResponse {
   items: CrmContactEntity[];
-  currentPage: number;
-  totalItems: number;
-  totalPages: number;
-}
-
-export interface CrmContactLookupListResponse {
-  items: CrmContactLookupItem[];
   currentPage: number;
   totalItems: number;
   totalPages: number;
@@ -224,7 +69,7 @@ export interface CrmTaskListResponse {
 }
 
 export interface CrmTaskTypeListResponse {
-  taskTypes: CrmTaskType[];
+  taskTypes: CrmTaskTypeEntity[];
 }
 
 export interface CrmTaskCompletedListResponse {
